fix styling, remove old blog page, add prismjs language supports
