+++ package.json
@@ -17,13 +17,15 @@
     "@nuxtjs/pwa": "^3.3.5",
     "core-js": "^3.16.1",
     "nuxt": "^2.15.8",
-    "prism-themes": "^1.8.0"
+    "prism-themes": "^1.8.0",
+    "prismjs": "^1.24.1"
   },
   "devDependencies": {
     "@nuxt/types": "^2.15.8",
     "@nuxt/typescript-build": "^2.1.0",
     "@nuxtjs/eslint-config-typescript": "^6.0.1",
     "@nuxtjs/eslint-module": "^3.0.2",
+    "@types/prismjs": "^1.16.6",
     "@vue/test-utils": "^1.2.2",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "^10.1.0",
